follow doppar doc convension:

+++ src/Phaseolies/Database/Query/Builder.php
@@ -6,6 +6,9 @@ use Phaseolies\Support\Facades\DB;
 
 class Builder
 {
+    /**
+     * @var string
+     */
     protected string $table;
 
     public function __construct(string $table)
@@ -15,11 +18,13 @@ class Builder
 
     /**
      * Truncate the table
+     * 
      * @return int
      */
     public function truncate(bool $resetAutoIncrement = true): int
     {
         $sql = "TRUNCATE TABLE {$this->table}";
+
         if (!$resetAutoIncrement) {
             $sql = "DELETE FROM {$this->table}";
         }
@@ -28,24 +33,26 @@ class Builder
             throw new \RuntimeException("Table {$this->table} does not exist");
         }
 
-        return DB::execute($sql);
+        return (int) DB::execute($sql);
     }
 
     /**
      * Delete all records from the table
+     * 
      * @return int
      */
     public function delete(): int
     {
-        return DB::execute("DELETE FROM {$this->table}");
+        return (int) DB::execute("DELETE FROM {$this->table}");
     }
 
     /**
      * Drop table from database
+     * 
      * @return int
      */
     public function drop(): int
     {
-        return DB::execute("DROP TABLE {$this->table}");
+        return (int) DB::execute("DROP TABLE {$this->table}");
     }
 }
